api2.0/requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
cryptography==41.0.2
docopt==0.6.2
ecdsa==0.18.0
fastapi==0.100.0
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pipreqs==0.4.13
pluggy==1.2.0
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.2
pydantic_core==2.1.2
PyJWT==2.8.0
pytest==7.4.0
pytest-httpx==0.22.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==4.6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy~=1.4.50
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
watchgod==0.8.2
Werkzeug==2.3.6
yarg==0.1.9

databases~=0.8.0
pip~=22.3.1
wheel~=0.38.4
exceptiongroup~=1.2.0
setuptools~=65.5.1